Fixes fakeArgs use, allows nested Whens

+++ WaffleTest/AssertionExtensions/FloatAssertions.cs
@@ -4,14 +4,14 @@ namespace WaffleTest.AssertionExtensions
 {
     public static class FloatAssertions
     {
-        public static Assertion MustBeInfinite(this ResultContext<float> result)
+        public static Assertion MustBeInfinite(this ResultContext<double> result)
         {
-            return result.Must(v => new Assertion(float.IsInfinity(v), string.Format("Expected +/-Infinity but got '{0}'", v)));
+            return result.Must(v => new Assertion(double.IsInfinity(v), string.Format("Expected +/-Infinity but got '{0}'", v)));
         }
 
-        public static Assertion MustBeNaN(this ResultContext<float> result)
+        public static Assertion MustBeNaN(this ResultContext<double> result)
         {
-            return result.Must(v => new Assertion(float.IsNaN(v), string.Format("Expected NaN but got '{0}'", v)));
+            return result.Must(v => new Assertion(double.IsNaN(v), string.Format("Expected NaN but got '{0}'", v)));
         }
     }
 }
